% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.vector_crossedUpAndDown.R
\name{predict.vector_crossedUpAndDown}
\alias{predict.vector_crossedUpAndDown}
\title{predict.vector_crossedUpAndDown}
\usage{
\method{predict}{vector_crossedUpAndDown}(object, vec)
}
\arguments{
\item{object}{the params from the fitted.}

\item{vec}{the vector to be applied}
}
\value{
a character vector
}
\description{
predict method for the class 'vector_crossedUpAndDown'.
}
