% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.vector_quantileSimple.R
\name{predict.vector_quantileSimple}
\alias{predict.vector_quantileSimple}
\title{predict.vector_quantileSimple}
\usage{
\method{predict}{vector_quantileSimple}(object, vec, fixLowerThanFirst = TRUE, fixHigherThanLast = TRUE)
}
\arguments{
\item{object}{the fitted model data.}

\item{vec}{a numeric vector.}

\item{fixLowerThanFirst}{if values are lower than the first quantile (supose to be the lowest)
that values will be assigned to that quantile.}

\item{fixHigherThanLast}{if values are higher than the last quantile (supose to be the highest)
that values will be assigned the that quantile.}
}
\value{
a character vector (or not).
}
\description{
the prediction method for the vector_quantileSimple class.
}
