% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_asNearestValue.R
\name{vector_asNearestValue}
\alias{vector_asNearestValue}
\title{vector_asNearestValue}
\usage{
vector_asNearestValue(vec, values)
}
\arguments{
\item{vec}{a numeric vector.}

\item{values}{a numeric vector with the values to be vec belong to.}
}
\value{
a numeric vector.
}
\description{
this will set the values of the input to the nearest value of the argument.
}
