% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_crossedUpAndDown.R
\name{vector_crossedUpAndDown}
\alias{vector_crossedUpAndDown}
\title{vector_crossedUpAndDown}
\usage{
vector_crossedUpAndDown(vec, crosses, crosses.names, tPlusX = 1)
}
\arguments{
\item{vec}{a numeric vector.}

\item{crosses}{a numeric vector with values which crosses will be used}

\item{crosses.names}{a list with vectors, each index of the list is a character vector with
the first position is the name to be applied to values which are coming from bellow to above the
crosses value.
example: list( c('crossUp20','crossDn20'), c('crossUp80','crossDn80') )}

\item{tPlusX}{is the lag to identify crosses.}
}
\value{
a list to be passed to the predict method.
}
\description{
will identify values which vector has crossed the especific points.
}
