% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_quantileSimple.R
\name{vector_quantileSimple}
\alias{vector_quantileSimple}
\title{vector_quantileSimple}
\usage{
vector_quantileSimple(vec, quants, ...)
}
\arguments{
\item{vec}{a numeric vector.}

\item{quants}{the split values for the quantilization, must be velues between 0 and 1.
tip: you can use the 'seq' function to generate this.}

\item{...}{options to be passed to stats::quantile (na.rm=TRUE is passed by default).}
}
\value{
a list with params to be passed to predict method.
}
\description{
does a simple quantilization (quantile classification) of the input vector.
}
