% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_quantileSimple.R
\name{vector_quantileSimple}
\alias{vector_quantileSimple}
\title{vector_quantileSimple}
\usage{
vector_quantileSimple(
  vec,
  quants,
  quants.names = (quants * 100) \%>\% as.character() \%>\% paste0(., "\%"),
  fixLowerThanFirst = TRUE,
  fixHigherThanLast = TRUE,
  ...
)
}
\arguments{
\item{vec}{a numeric vector.}

\item{quants}{the split values for the quantilization, must be velues between 0 and 1.
tip: you can use the 'seq' function to generate this.}

\item{quants.names}{the names to be applied to the quantiles.}

\item{fixLowerThanFirst}{if values are lower than the first quantile (supose to be the lowest)
that values will be assigned to that quantile.}

\item{fixHigherThanLast}{if values are higher than the last quantile (supose to be the highest)
that values will be assigned the that quantile.}

\item{...}{options to be passed to stats::quantile (na.rm=TRUE is passed by default).}
}
\value{
a character vector.
}
\description{
does a simple quantilization of the input vector.
}
